requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-filter==23.5
dj-database-url==2.1.0
